// src/app/pets/models/pet-details.ts
export interface petDetails {
  //define variable here
  petId: number;
  petName: string;
  petAge: string;
  petStatus: boolean;
  parentFName: string;
  parentLName: string;
  parentEmailId: string;
  parentPhoneNumber: string;
  breed: string;
  petGender: string;
  biography: string;
  city: string;
  state: string;
  country: string;
  pincode: number;
  avatar: string;
}